requirements.txt
streamlit==1.35.0
transformers==4.41.2
torch>=2.0.0
pdfminer.six==20221105
